pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "carbon-sequestration-estimator"
version = "1.0.0"
description = "High-resolution satellite imagery analysis for carbon sequestration quantification."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Aniket Shinde" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: GIS",
]
dependencies = [
    "streamlit>=1.32.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "opencv-python-headless>=4.8.0",
    "matplotlib>=3.7.0",
    "Pillow>=10.0.0",
    "rasterio>=1.3.0",
    "detectree>=0.4.4",
    "scikit-image>=0.21.0",
]

[project.urls]
Homepage = "https://github.com/aniketqxp/carbon-sequestration-index"
Issues = "https://github.com/aniketqxp/carbon-sequestration-index/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "src.*"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88
